MAJOR QOL CHANGES
Flyby has always struggled with input handing, and today, with the help of Microsoft Copilot, we've moved away from fmt.Scan (so long, fella!) and moved to "bufio" (no idea what that is) but it handles spaces better.
Bets are now strings that get converted to integers, so we can have things that happen when you put in certain text (instead of it just bugging out)
Also, while rolling and writing, key presses will no longer register, meaning it's impossible double tap enter and mess up the whole flyby screen. Exciting!
As usual, code is awful.