Added two characters (Bartender and drug dealer)
You can now smoke weed, drink beer, kill both the bartender and drug dealer, as well as commit suicide.
Added new stats on the win page, but only if you do one of the associated actions.
By the way, I've added a debug command. type "redraw" in the bet to redraw the machine. Useful if it breaks for whatever reason.